Update fps code

+++ src/settings.h
@@ -73,7 +73,7 @@ typedef struct utox_settings {
     uint32_t window_baseline;
 
     bool    window_maximized;
-    uint32_t video_fps;
+    uint8_t video_fps;
 } SETTINGS;
 
 extern SETTINGS settings;
@@ -122,7 +122,7 @@ typedef struct utox_save {
 
     uint16_t language;
 
-    uint32_t video_fps;
+    uint8_t  video_fps;
     uint16_t unused[25];
     uint8_t  proxy_ip[];
 } UTOX_SAVE;
